Zeichen-Vergleich bei Buchstaben muss mit (unsigned) byte statt char durchgeführt werden!

+++ source/MovieCutter.c
@@ -6822,7 +6822,7 @@ bool MovieCutterRenameFile(void)
 
 #if STACKTRACE == TRUE
   TAP_PrintNet("DEBUG: VOR der Umwandlung in ISO:\n");
-  TAP_PrintNet("DEBUG: PlaybackName = [%x] '%s'\n", (PlaybackName[0] >= 0x20 ? 0 : PlaybackName[0]), PlaybackName);
+  TAP_PrintNet("DEBUG: PlaybackName = [%x] '%s'\n", ((byte)PlaybackName[0] >= 0x20 ? 0 : PlaybackName[0]), PlaybackName);
 #endif
 
   ExtensionStart = strrchr(PlaybackName, '.');
@@ -6836,7 +6836,7 @@ bool MovieCutterRenameFile(void)
 
 #if STACKTRACE == TRUE
   TAP_PrintNet("DEBUG: NACH der Umwandlung in ISO:\n");
-  TAP_PrintNet("DEBUG: TempNameISO  = [%x] '%s'\n", (TempNameISO[0] >= 0x20 ? 0 : TempNameISO[0]), TempNameISO);
+  TAP_PrintNet("DEBUG: TempNameISO  = [%x] '%s'\n", ((byte)TempNameISO[0] >= 0x20 ? 0 : TempNameISO[0]), TempNameISO);
 #endif
 
   while (!ret)
@@ -6856,7 +6856,7 @@ bool MovieCutterRenameFile(void)
     OSDRedrawEverything();
 
     // Eingabe überprüfen und ggf. wieder in UTF-8 umwandeln
-    if ((TempNameISO[0] >= 0x20) || (TempNameISO[0] && TempNameISO[1]))
+    if (((byte)TempNameISO[0] >= 0x20) || (TempNameISO[0] && TempNameISO[1]))
     {
       // Wenn Name ungültig, Meldung anzeigen
       if (strchr(TempNameISO, '/'))
@@ -6869,10 +6869,10 @@ bool MovieCutterRenameFile(void)
 
 #if STACKTRACE == TRUE
   TAP_PrintNet("DEBUG: VOR Rück-Umwandlung in UTF-8:\n");
-  TAP_PrintNet("DEBUG: TempNameISO  = [%x] '%s'\n", (TempNameISO[0] >= 0x20 ? 0 : TempNameISO[0]), TempNameISO);
+  TAP_PrintNet("DEBUG: TempNameISO  = [%x] '%s'\n", ((byte)TempNameISO[0] >= 0x20 ? 0 : TempNameISO[0]), TempNameISO);
 #endif
 
-        TempNameFull = (TempNameISO[0] >= 0x20) ? &TempNameBuffer[1] : TempNameBuffer;
+        TempNameFull = ((byte)TempNameISO[0] >= 0x20) ? &TempNameBuffer[1] : TempNameBuffer;
         strcpy(TempNameFull, TempNameISO);
         if (isUTFToppy() && TempNameISO[0] >= 0x15)
           ConvertUTFStr(NewName, TempNameFull, sizeof(NewName) - strlen(ExtensionStart) - 4, TRUE);
@@ -6881,15 +6881,15 @@ bool MovieCutterRenameFile(void)
         strcat(NewName, ExtensionStart);
         strcat(TempNameFull, ExtensionStart);
 
-        PlayNameNoTypeChar = (PlaybackName[0] >= 0x20) ? PlaybackName : &PlaybackName[1];
-        TempNameNoTypeChar = &TempNameBuffer[1];  // (TempNameFull[0] >= 0x20) ? TempNameFull : &TempNameFull[1];
-//        TempNameWithTypeChar = (TempNameFull[0] >= 0x20) ? TempNameBuffer : TempNameFull;
+        PlayNameNoTypeChar = ((byte)PlaybackName[0] >= 0x20) ? PlaybackName : &PlaybackName[1];
+        TempNameNoTypeChar = &TempNameBuffer[1];  // ((byte)TempNameFull[0] >= 0x20) ? TempNameFull : &TempNameFull[1];
+//        TempNameWithTypeChar = ((byte)TempNameFull[0] >= 0x20) ? TempNameBuffer : TempNameFull;
 
 #if STACKTRACE == TRUE
   TAP_PrintNet("DEBUG: NACH der Rück-Umwandlung in UTF-8:\n");
-  TAP_PrintNet("DEBUG: TempNameBuf  = [%x] '%s'\n", (TempNameBuffer[0] >= 0x20 ? 0 : TempNameBuffer[0]), TempNameBuffer);
-  TAP_PrintNet("DEBUG: TempNameFull = [%x] '%s'\n", (TempNameFull[0]   >= 0x20 ? 0 : TempNameFull[0]),   TempNameFull);
-  TAP_PrintNet("DEBUG: NewName      = [%x] '%s'\n", (NewName[0]        >= 0x20 ? 0 : NewName[0]),        NewName);
+  TAP_PrintNet("DEBUG: TempNameBuf  = [%x] '%s'\n", ((byte)TempNameBuffer[0] >= 0x20 ? 0 : TempNameBuffer[0]), TempNameBuffer);
+  TAP_PrintNet("DEBUG: TempNameFull = [%x] '%s'\n", ((byte)TempNameFull[0]   >= 0x20 ? 0 : TempNameFull[0]),   TempNameFull);
+  TAP_PrintNet("DEBUG: NewName      = [%x] '%s'\n", ((byte)NewName[0]        >= 0x20 ? 0 : NewName[0]),        NewName);
 #endif
 
         // Wenn neuer Name nicht leer, aber verändert
